Pagecontroll, routing, font
The font style was changed to Inter.
The basic routes were established and everything was organized in corresponding folders.
The bottomAppBar was further developed by integrating the PageView and connecting the buttons to the corresponding pages, allowing the control of the view by swiping or clicking on the buttons. Animation for the FAB was implemented and connected with the performance_screen.